course_project/todo-backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Todo Backend API Service"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.41",
    "asyncpg>=0.30.0",
    "pytest-asyncio>=1.1.0",
    "greenlet>=3.2.3",
    "nats-py>=2.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "pyyaml>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["E501", "B008"]  # B008: FastAPI Depends() in defaults is standard pattern

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Use spaces around operators
indent-style = "space"
# Skip magic trailing commas for single-element tuples
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
known-first-party = ["src"]
# Force single line imports for better readability
force-single-line = false
# Compatible with black formatting
combine-as-imports = true
split-on-trailing-comma = true
